Assignment.Clients.BlockWebsocket (assignment v0.1.0)

Websocket client to communicate with blocknative websocket api server. It uses the websockex library and is started supervised as a genserver. handle_frame callback handles incoming message from the websocket server. How it works:

Connect to the blocknative websocket server.

Send an initialization request where the api key is verified.

Send request to get transaction status of pending transactions.

Send the response(pending/confirmed) to the user websocket server(TxWebsocket) using pubsub.

Send the response to a slack webhook as well.

The client automatically connects and sends initialization request to the blocknative server when the application starts.

handle_message/1 is a private function that handles the incoming messages from blocknative. An incoming message can be a successful initialization event or a transaction status event. Types of messages sent to blocknative: Initialization payload:

{
  "timeStamp":"2021-01-11T06:21:40.197Z",
  "dappId":"xyz",
  "version":"1",
  "blockchain":{
      "system":"ethereum",
      "network":"main"
  },
  "categoryCode":"initialize",
  "eventCode":"checkDappId"
}

Initialization response payload:

{
  "connectionId": "C4-e78f1dcd-920e-48ea-bff8-d39f88719151",
  "serverVersion": "0.122.2",
  "showUX": true,
  "status": "ok",
  "version": 0
}

Get status of a transaction payload:

{
    "timeStamp": "2021-01-11T06:21:40.197Z",
    "dappId": "xyz",
    "version": "1",
    "blockchain": {
        "system": "ethereum",
        "network": "main"
    },
    "categoryCode": "accountAddress",
    "eventCode": "txSent",
    "transaction": {
        "hash": "0x794b11732e221d1c38276c79650a3fdb4594697db6e030066167e0403b1ef369",
        "id": "0x794b11732e221d1c38276c79650a3fdb4594697db6e030066167e0403b1ef369"
    }
}

When transaction is confirmed we get a message with event code: txConfirmed

Functions

get_tx_status(pid, tx_id)

This method creates a request payload for getting transaction status and sends it to blocknative websocket server.

send_message(client, message)

Sends message to the blocknative websocket server.