Assignment.TxWebsocket (assignment v0.1.0)

This is the websocket server that communicates with user about their transaction status. This is a raw websocket server written over Phoenix.Socket.Transport Working:

The front-end connects with websocket by passing user_id.

The websocket server accepts the connection request and subscribes the unique user id to a pub sub topic.

The server then stores the tx_id and user_id mapping in ets and calls get_tx_status/2 of the blocknative websocket client to send request for getting transaction status.

The blocknative websocket client uses the ets tx_id->user_id mapping to broadcast the tx_id status back to the user_id topic. The websocket server doesn't handle authentication for now.

Functions

connect(state)

Callback implementation for Phoenix.Socket.Transport.connect/1.

handle_in(arg, state)

Callback implementation for Phoenix.Socket.Transport.handle_in/2.

handle_info(message, state)

Callback implementation for Phoenix.Socket.Transport.handle_info/2.

init(state)

Callback implementation for Phoenix.Socket.Transport.init/1.

terminate(reason, state)

Callback implementation for Phoenix.Socket.Transport.terminate/2.